Make command handler actually work

+++ src/types/chat.rs
@@ -2,10 +2,10 @@ use crate::types::{ChatLocation, ChatPermissions, ChatPhoto, Location, Message, 
 
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Deserialize, Serialize)]
+#[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct Chat {
     /// Unique identifier for this chat. This number may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier.
-    pub id: i32,
+    pub id: i64,
     /// Type of chat, can be either "private", "group", "supergroup" or "channel"
     pub r#type: String,
     /// Optional. Title, for supergroups, channels and group chats
@@ -35,7 +35,7 @@ pub struct Chat {
     /// Optional. True, if the bot can change the group sticker set. Returned only in getChat.
     pub can_set_sticker_set: Option<bool>,
     /// Optional. Unique identifier for the linked chat, i.e. the discussion group identifier for a channel and vice versa; for supergroups and channel chats. This identifier may be greater than 32 bits and some programming languages may have difficulty/silent defects in interpreting it. But it is smaller than 52 bits, so a signed 64 bit integer or double-precision float type are safe for storing this identifier. Returned only in getChat.
-    pub linked_chat_id: Option<i32>,
+    pub linked_chat_id: Option<i64>,
     /// Optional. For supergroups, the location to which the supergroup is connected. Returned only in getChat.
     pub location: Option<ChatLocation>,
 }
